requirements.txt
Flask==1.1.2
discord.py==1.7.3
urllib3==1.26.7
requests==2.25.1
discord==1.7.3
python-dotenv==0.19.2
python_dateutil==2.8.2
buttons==0.1.9